+++ CHANGELOG.md
@@ -54,6 +54,7 @@ The major version bump is due to upping the required PHP version from `8.5` to `
 	- `ImageInterface::toDataUri()`
 	- `ImageInterface::toStream()`
 	- `ImageInterface::getMimeType()`
+* Added operation `Pipeline` class to the pixel library.
 * Added dotenv loader.
 * Added `Signer::validateOrThrow()` method.
 * Added support for all new Redis commands added in Redis 8.2, 8.4, 8.6, 8.8 and 8.10 to the Redis client.